extract user tags tests into a separate file
